Some fixes in the code

+++ XS3GNR_HFT_2021221.Logic/FacultyLogic.cs
@@ -42,6 +42,10 @@ namespace XS3GNR_HFT_2021221.Logic
 
         public void Delete(int id)
         {
+            if (id < 1)
+            {
+                throw new ArgumentException("Id cannot be negative");
+            }
             facultyRepo.Delete(id);
         }
